backend/roadvisor/models/requirements.txt
google-cloud-aiplatform==1.0.1
google-cloud-storage==1.42.0
protobuf==3.17.3
numpy==1.21.2
Pillow==8.3.1
